print usage on invalid CLI arguments
This improves the usability by printing the command usage if the user
provides the wrong number of arguments or invalid flags to a subcommand.

Previously, only a short message like

    accepts 1 arg(s), received 0

or

    unknown flag: --foobar

was shown.

Now the full usage with example will be printed above the error in case
argument parsing fails.

Please note that command usage is not printed if the user provided valid
arguments but the command fails due to other reasons (e.g. usage queries
for a skeleton that does not exist).