add --overwrite flag to `project create` command (#43)
This causes a change to the default behaviour. Previously, files already
existing in the target dir were overwritten. It is safer to retain
those files by default and only overwrite them if the user explicitly
passes the `--overwrite` flag to the `project create` command.

See the updated usage examples in `pkg/cmd/project/create.go`.