remove email from config
We should not expect users to enter an email address into a random tool
if the email address is not required for the tool to function correctly.
In kickoff it was only used in the `[fullname]` field of license texts,
which is not required. The project owner is sufficient there.

Therefore, we completely remove support for project email from the code
base. Skeleton creators can still opts for requiring `email` via a
custom user-defined `.Values` field.