move pkg/ to internal/
In the current state of the project we should not encourage people to
depend on one of our packages as the APIs are still changing randomly.

It is better to declare everything as `internal` now and then move
packages into `pkg` where we feel that the APIs will stay stable and
that might have a value being used outside of kickoff.