add --repository flag
This adds the --repository flag which can be used to filter for
skeletons only from a specific repository. The flag can be specified
multiple times. The makes the (badly designed) --repositories flag
obsolete and it is thus removed.

To ease testing of commands that accept a config file, the `testutil`
package was introduced which contains a `ConfigFileBuilder` for building
temporary config files on the fly.

Usages of hardcoded config files in tests were replaced with the
`ConfigFileBuilder` where possible.