add http response cache
This adds an http response cache that will transparently cache http
responses that are cacheable. The cache is wrapped by another
`http.RoundTripper` that will attempt to serve potentially stale
responses from cache if there is not internet connectivity, to make
offline usage of kickoff features that fetch data from third parties
possible to a certain degree, that is: if the user already requested a
resource previously, it will log a warning and serve a potentially stale
cached copy if the user is offline.