feat(project): better project creation output
This replaces the console logger that was used to generate project
creation output with something that is better suited for the job.

The change made it necessary to rewrite the whole project creation logic
to avoid falling back to ugly hacks. The new project options pattern
allows to easily add more options in the future if necessary.