fix(logging)!: distinguish between logs and output (#75)
The logger should only be used to for additional information that is
useful for debugging. Normal output should just go to stdout.

Changes log stmt to stdout messages where appropriate.

Also removed the `--verbose` flag and replaces it with the `--log-level`
flag to allow more fine grained control over the log output.