feat(logging): switch to logrus
With the switch to logrus we have file and function info available in
the log output which makes #69 obsolete as we do not need to enrich logs
with the component as this can be inferred from the import path.

The new log format also better distinguishes log output from normal
output.

Closes #69