feat: use repo's default branch if revision is omitted
Until now, the `master` branch was picked if no revision was specified.
Since lots of git hosting platforms switched the default to `main` it
makes sense to not default to master anymore.

Instead, just clone the repo (which will checkout the default branch)
and do not attempt to checkout a revision.