feat!: improve ux of the project creation command (#118)
This changes the project creation command from:

```
kickoff project create <skeleton-name> <dir>
```

to

```
kickoff project create <project-name> <skeleton-name> [<skeleton-name>...]
```

This is a breaking change but helps with autocompletion of skeleton
names. Furthermore it now accepts a project name instead of a directory
as argument. The project will be created in `$PWD/<project-name>` by
default but this can be overridden via the `--dir` flag. As a
consequence the `--name` flag was removed.