refactor: remove non-panicking Expand/Collapse
These were only used internally anyways and kickoff has no real use for
them. Panicking is the best option we have if the homedir cannot be
detected and kickoff relies on it and would not work without it
properly.

As a consequence `MustCollapse` and `MustExpand` were renamed to
`Collapse` and `Expand` to make the API cleaner.