feat: move to GitHub Gitignores API
Replaces usage of the https://www.toptal.com/developers/gitignore API
with the GitHub Gitignores API.

While this change required a little more code to get to the same
behaviour as the old API showed, it has some advantages:

- Reduces kickoff's third-party API dependencies to one.
- Gitignore templates can finally be cached properly. The toptal API
  marked requests as uncacheable in their responses. This was very
  unfortunate as there was a noticeable delay in any kickoff operation
  involving gitignores. GitHub set proper cache headers which should
  speed things up considerably.

The change will decrease the number of available gitignore templates
from 522 to 127 but I think that less is more and since gitignore
templates can be composed (e.g. "go,python,...") this should be enough.