feat: interactive project creation (#131)
Interactive mode is used if one or more of the following conditions
apply:

- `-i/--interactive` flag is explicitly provided
- missing skeleton name (`len(args) < 2`)
- missing project and skeleton name (`len(args) == 0`)

In interactive mode, the user is only prompted for information that
he/she did not explicitly set via flags already. E.g. if `--license` was
provided interactive license selection is skipped.

The feature may still have bugs. An extensive test suite for the
`project create` command will be added soon to catch these. For now
there are no tests as stubbing infrastructure for testing of interactive
mode needs to be created first.

Closes https://github.com/martinohmann/kickoff/issues/81